GasLog
I wanted to graph the fuel efficiency of my 2000 Toyota Corolla over a
few months. On top of that I wanted to see what kind of stuff I could
learn and do with this kind of data. In particular, I want to be able
to visualize this data in the browser.